test_autodoc: Separate testdata (python objects) and testcases
