+++ tests/roots/test-root/autodoc.txt
@@ -5,7 +5,7 @@ Just testing a few autodoc possibilities...
 
 .. automodule:: util
 
-.. automodule:: test_autodoc
+.. automodule:: autodoc_target
    :members:
 
 .. autofunction:: function
@@ -34,7 +34,7 @@ Just testing a few autodoc possibilities...
    .. autoclass:: MarkupError
 
 
-.. currentmodule:: test_autodoc
+.. currentmodule:: autodoc_target
 
 .. autoclass:: InstAttCls
    :members:
