+++ tests/test_ext_coverage.py
@@ -21,9 +21,9 @@ def test_build(app, status, warning):
     py_undoc = (app.outdir / 'python.txt').text()
     assert py_undoc.startswith('Undocumented Python objects\n'
                                '===========================\n')
-    assert 'test_autodoc\n------------\n' in py_undoc
+    assert 'autodoc_target\n--------------\n' in py_undoc
     assert ' * Class -- missing methods:\n' in py_undoc
-    assert ' * process_docstring\n' in py_undoc
+    assert ' * raises\n' in py_undoc
     assert ' * function\n' not in py_undoc  # these two are documented
     assert ' * Class\n' not in py_undoc     # in autodoc.txt
 
@@ -40,9 +40,9 @@ def test_build(app, status, warning):
     # the key is the full path to the header file, which isn't testable
     assert list(undoc_c.values())[0] == set([('function', 'Py_SphinxTest')])
 
-    assert 'test_autodoc' in undoc_py
-    assert 'funcs' in undoc_py['test_autodoc']
-    assert 'process_docstring' in undoc_py['test_autodoc']['funcs']
-    assert 'classes' in undoc_py['test_autodoc']
-    assert 'Class' in undoc_py['test_autodoc']['classes']
-    assert 'undocmeth' in undoc_py['test_autodoc']['classes']['Class']
+    assert 'autodoc_target' in undoc_py
+    assert 'funcs' in undoc_py['autodoc_target']
+    assert 'raises' in undoc_py['autodoc_target']['funcs']
+    assert 'classes' in undoc_py['autodoc_target']
+    assert 'Class' in undoc_py['autodoc_target']['classes']
+    assert 'undocmeth' in undoc_py['autodoc_target']['classes']['Class']
